backend/requirements.txt
fastapi==0.99.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
httpx==0.24.0
pydantic==2.6.0
